Move validator test to appropriate package
